something… probably fixing compile warnings. not sure

+++ fib.c
@@ -5,7 +5,7 @@
 
 static unsigned long long fib_num(int n) {
     unsigned long long a = 0, b = 1;
-    for (int i = 31; i >= 0; i--) {
+    for (long i = 31; i >= 0; i--) {
 	unsigned long long d = a * (b * 2 - a);
 	unsigned long long e = a * a + b * b;
 	a = d;
@@ -26,13 +26,13 @@ void fib(int fd, int direction) {
     struct led leds[NUM_LEDS];
     struct hsv hsv_tuples[N];
     struct rgb a;
-    for(int i=0; i<N; ++i) {
+    for(size_t i=0; i<N; ++i) {
         hsv_tuples[i].h = i/(double) N*360.0;
         hsv_tuples[i].s = 0.9;
         hsv_tuples[i].v = 0.7;
     }
-    for(int x=0; x<N; ++x) {
-    	for(int i=fib_num(x); i<fib_num(x+1) && i<NUM_LEDS; ++i) {
+    for(size_t x=0; x<N; ++x) {
+        for(size_t i=fib_num(x); i<fib_num(x+1) && i<NUM_LEDS; ++i) {
     	    a = hsv2rgb(hsv_tuples[x]);
     	    leds[i].r = (uint8_t) (a.r*factor);
     	    leds[i].g = (uint8_t) (a.g*factor);
